Article summary web tests (#332)
* lint

* added tests for web

* removed unused fixture in article-summary-stories

* updated snapshots
