+++ packages/article-summary/__tests__/article-summary.native.test.js
@@ -1,0 +1,7 @@
+/* eslint-env jest */
+
+import "react-native";
+import ArticleSummary from "../article-summary";
+import test from "./article-summary-helper";
+
+test(ArticleSummary);
