+++ packages/article-summary/__tests__/article-summary.web.test.js
@@ -1,0 +1,7 @@
+/* eslint-env jest */
+
+import "react-native-web";
+import ArticleSummary from "../article-summary";
+import test from "./article-summary-helper";
+
+test(ArticleSummary);
