+++ packages/article-summary/jest.config.json
@@ -1,0 +1,8 @@
+{
+  "preset": "react-native",
+  "rootDir": "../../",
+  "transformIgnorePatterns": ["node_modules/(?!@times-components)/"],
+  "testMatch": [
+    "<rootDir>/packages/article-summary/__tests__/**!(.web).test.js"
+  ]
+}
