+++ packages/article-summary/jest.config.web.json
@@ -1,0 +1,11 @@
+{
+  "preset": "react-native",
+  "rootDir": "../../",
+  "transformIgnorePatterns": ["node_modules/(?!@times-components)/"],
+  "testMatch": ["<rootDir>/packages/article-summary/__tests__/**.web.test.js"],
+  "moduleNameMapper": {
+    "react-native": "react-native-web"
+  },
+  "moduleFileExtensions": ["web.js", "js", "json"],
+  "testEnvironment": "jsdom"
+}
