+++ packages/article-summary/package.json
@@ -4,8 +4,10 @@
   "main": "article-summary",
   "scripts": {
     "flow": "node_modules/flow-bin/cli.js",
-    "test:watch": "jest --bail --verbose --watchAll",
-    "test": "jest --bail --ci --coverage",
+    "test:watch": "npm run test:native -- --bail --verbose --watchAll",
+    "test:native": "jest --config='./jest.config.json'",
+    "test:web": "jest --config='./jest.config.web.json'",
+    "test": "npm run test:native -- --ci --bail && npm run test:web -- --ci --bail --coverage",
     "prettier:diff": "prettier --list-different '**/*.+(js|json)'",
     "lint": "eslint . && npm run prettier:diff"
   },
